requirements.txt
Django==5.2.7
mysqlclient
python-dotenv
Pillow
